+++ src/Plugin.cs
@@ -42,6 +42,7 @@ public class Plugin : BaseUnityPlugin
         _harmony = new Harmony(MyPluginInfo.PLUGIN_GUID);
         _harmony.PatchAll(typeof(StartOfRoundPatches));
         _harmony.PatchAll(typeof(SetPlanetsWeatherPatch));
+        _harmony.PatchAll(typeof(RoundManagerPatches));
         _harmony.PatchAll(typeof(TimeOfDayPatches));
 
         WeatherUnlocker.Subscribe();